Test cases pass

+++ src/main/java/ampcontrol/model/training/model/vertex/ElementWiseVertexLatest.java
@@ -63,7 +63,7 @@ public class ElementWiseVertexLatest extends GraphVertex {
     }
 
     @Override
-    public int numParams(boolean backprop) {
+    public long numParams(boolean backprop) {
         return 0;
     }
